requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic
pytest
pydantic[email]
python-multipart
httpx